imageList の更新に関する機能を、一つのファイルにまとめた。
