+++ src/generateImageList.js
@@ -1,8 +1,24 @@
-import specifyShowState from './specifyShowState';
+export const SPECIFY_SHOW_STATE = 'specifyShowState';
 
-export default function generateImageList(type, data, currentState) {
+function specifyShowState(data, currentState) {
+  const {target, isShow} = data;
+  let targetData;
+  let targetIndex;
+  currentState.forEach((elem, index) => {
+    if (elem.id === target) {
+      targetData = elem;
+      targetIndex = index;
+    }
+  });
+  const updatedData = Object.assign({}, targetData, {isShow});
+  const updatedState = Object.assign([], currentState);
+  updatedState.splice(targetIndex, 1, updatedData);
+  return updatedState;
+}
+
+export function generateImageList(type, data, currentState) {
   switch (type) {
-    case 'specifyShowState':
+    case SPECIFY_SHOW_STATE:
       return specifyShowState(data, currentState);
     default:
       throw new Error('This type is not found.');
